Committing update to trunk.  Revision updated to was: 27490.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2007/collada@27492 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/decal/decal.cpp
@@ -163,7 +163,7 @@ void csDecal::BeginMesh(iMeshWrapper * mesh)
 void csDecal::AddStaticPoly(const csPoly3D & p)
 {
   size_t a;
-  CS::TriangleT<size_t> tri;
+  CS::TriangleT<int> tri;
 
   if (!currMesh)
       return;
